backend/requirements.txt
flask==3.0.3
flask-socketio==5.4.1
eventlet==0.33.3
mediapipe==0.10.14
opencv-python==4.10.0.84
torch==2.3.1
torchvision==0.18.1
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
tqdm==4.66.4
